+++ src/ifs/wl_surface/x_surface/xwindow.rs
@@ -302,7 +302,7 @@ impl Xwindow {
             Change::Map => {
                 self.data.state.map_tiled(self.clone());
                 if self.data.info.fullscreen.get() {
-                    self.clone().tl_set_fullscreen(true);
+                    self.clone().tl_set_fullscreen(true, None);
                 }
                 self.data.title_changed();
             }